langchain-nunsys-back/requirements.txt
# openai
# numpy==1.24.1
# fastapi==0.89.1
pydantic==1.10.2
# torch==1.13.1
langchain
# websockets==10.4
# tqdm==4.64.1
langchain_openai
langchain_chroma
langchain-community
chromadb
pypdf
pyowm
load_dotenv
langgraph